Added encoder's MHSA QK^T and /sqrt(8)

+++ asic/func_sim/include/Compute_Verification.hpp
@@ -70,11 +70,11 @@ static std::map<COMPUTE_VERIFICATION_STEP, StepVerifInfo> step_verif_info = {
 
 /*----- FUNCTION -----*/
 #if DISTRIBUTED_ARCH
-bool are_equal(float a, float b, uint16_t index, uint8_t id);
+bool are_equal(float a, float b, int32_t index, uint8_t id);
 void verify_result(RESULT_TYPE type, float result, float* input_data, uint16_t starting_addr, uint16_t len, uint8_t id, DATA_WIDTH data_width);
 void verify_layer_out(COMPUTE_VERIFICATION_STEP cim_state, uint8_t id, float* data, uint16_t starting_addr, DATA_WIDTH data_width);
 #elif CENTRALIZED_ARCH
-bool are_equal(float a, float b, uint16_t index);
+bool are_equal(float a, float b, int32_t index);
 void verify_result(RESULT_TYPE type, float result, float* input_data, uint16_t starting_addr, uint16_t len, DATA_WIDTH data_width);
 void verify_layer_out(COMPUTE_VERIFICATION_STEP cim_state, float* data, uint16_t starting_addr, uint16_t outside_dim_len, DATA_WIDTH data_width);
 #endif
